+++ Source/Rows/CheckRow.swift
@@ -23,12 +23,13 @@
 // THE SOFTWARE.
 
 import Foundation
+import UIKit
 
 // MARK: CheckCell
 
 public final class CheckCell: Cell<Bool>, CellType {
 
-    required public init(style: UITableViewCellStyle, reuseIdentifier: String?) {
+    required public init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
         super.init(style: style, reuseIdentifier: reuseIdentifier)
     }
 
@@ -36,7 +37,7 @@ public final class CheckCell: Cell<Bool>, CellType {
         super.init(coder: aDecoder)
     }
 
-    open override func update() {
+    public override func update() {
         super.update()
         accessoryType = row.value == true ? .checkmark : .none
         editingAccessoryType = accessoryType
@@ -49,13 +50,13 @@ public final class CheckCell: Cell<Bool>, CellType {
         }
     }
 
-    open override func setup() {
+    public override func setup() {
         super.setup()
         accessoryType = .checkmark
         editingAccessoryType = accessoryType
     }
 
-    open override func didSelect() {
+    public override func didSelect() {
         row.value = row.value ?? false ? false : true
         row.deselect()
         row.updateCell()