+++ Source/Rows/Common/DateInlineFieldRow.swift
@@ -23,10 +23,11 @@
 // THE SOFTWARE.
 
 import Foundation
+import UIKit
 
 open class DateInlineCell: Cell<Date>, CellType {
 
-    public required init(style: UITableViewCellStyle, reuseIdentifier: String?) {
+    public required init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
         super.init(style: style, reuseIdentifier: reuseIdentifier)
     }