+++ Source/Rows/Common/SelectorRow.swift
@@ -23,10 +23,11 @@
 // THE SOFTWARE.
 
 import Foundation
+import UIKit
 
 open class PushSelectorCell<T: Equatable> : Cell<T>, CellType {
 
-    required public init(style: UITableViewCellStyle, reuseIdentifier: String?) {
+    required public init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
         super.init(style: style, reuseIdentifier: reuseIdentifier)
     }