+++ Source/Rows/PickerInlineRow.swift
@@ -23,10 +23,11 @@
 // THE SOFTWARE.
 
 import Foundation
+import UIKit
 
 open class PickerInlineCell<T: Equatable> : Cell<T>, CellType {
 
-    required public init(style: UITableViewCellStyle, reuseIdentifier: String?) {
+    required public init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
         super.init(style: style, reuseIdentifier: reuseIdentifier)
     }
 
@@ -88,7 +89,7 @@ public final class PickerInlineRow<T> : _PickerInlineRow<T>, RowType, InlineRowT
     public func setupInlineRow(_ inlineRow: InlineRow) {
         inlineRow.options = self.options
         inlineRow.displayValueFor = self.displayValueFor
-        inlineRow.cell.height = { UITableViewAutomaticDimension }
+        inlineRow.cell.height = { UITableView.automaticDimension }
     }
 }
 
@@ -140,7 +141,7 @@ public final class DoublePickerInlineRow<A, B> : _DoublePickerInlineRow<A, B>, R
         inlineRow.firstOptions = firstOptions
         inlineRow.secondOptions = secondOptions
         inlineRow.displayValueFor = self.displayValueFor
-        inlineRow.cell.height = { UITableViewAutomaticDimension }
+        inlineRow.cell.height = { UITableView.automaticDimension }
     }
 }
 
@@ -196,6 +197,6 @@ public final class TriplePickerInlineRow<A, B, C> : _TriplePickerInlineRow<A, B,
         inlineRow.secondOptions = secondOptions
         inlineRow.thirdOptions = thirdOptions
         inlineRow.displayValueFor = self.displayValueFor
-        inlineRow.cell.height = { UITableViewAutomaticDimension }
+        inlineRow.cell.height = { UITableView.automaticDimension }
     }
 }