+++ Source/Rows/PickerRow.swift
@@ -23,6 +23,7 @@
 // THE SOFTWARE.
 
 import Foundation
+import UIKit
 
 // MARK: PickerCell
 
@@ -32,7 +33,7 @@ open class _PickerCell<T> : Cell<T>, CellType, UIPickerViewDataSource, UIPickerV
 
     fileprivate var pickerRow: _PickerRow<T>? { return row as? _PickerRow<T> }
 
-    public required init(style: UITableViewCellStyle, reuseIdentifier: String?) {
+    public required init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
         let pickerView = UIPickerView()
         self.picker = pickerView
         self.picker?.translatesAutoresizingMaskIntoConstraints = false
@@ -52,7 +53,7 @@ open class _PickerCell<T> : Cell<T>, CellType, UIPickerViewDataSource, UIPickerV
         super.setup()
         accessoryType = .none
         editingAccessoryType = .none
-        height = { UITableViewAutomaticDimension }
+        height = { UITableView.automaticDimension }
         picker.delegate = self
         picker.dataSource = self
     }
@@ -69,7 +70,7 @@ open class _PickerCell<T> : Cell<T>, CellType, UIPickerViewDataSource, UIPickerV
         picker?.dataSource = nil
     }
 
-    open var pickerTextAttributes: [NSAttributedStringKey: Any]?
+    open var pickerTextAttributes: [NSAttributedString.Key: Any]?
 
     open func numberOfComponents(in pickerView: UIPickerView) -> Int {
         return 1
@@ -103,7 +104,7 @@ open class PickerCell<T> : _PickerCell<T> where T: Equatable {
         super.init(coder: aDecoder)
     }
 
-    public required init(style: UITableViewCellStyle, reuseIdentifier: String?) {
+    public required init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
         super.init(style: style, reuseIdentifier: reuseIdentifier)
     }