+++ Source/Rows/TriplePickerRow.swift
@@ -7,6 +7,7 @@
 //
 
 import Foundation
+import UIKit
 
 public struct Tuple3<A: Equatable, B: Equatable, C: Equatable> {
     public let a: A
@@ -33,7 +34,7 @@ open class TriplePickerCell<A, B, C> : _PickerCell<Tuple3<A, B, C>> where A: Equ
 
     private var pickerRow: _TriplePickerRow<A, B, C>? { return row as? _TriplePickerRow<A, B, C> }
 
-    public required init(style: UITableViewCellStyle, reuseIdentifier: String?) {
+    public required init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
         super.init(style: style, reuseIdentifier: reuseIdentifier)
     }
 
@@ -89,9 +90,9 @@ open class TriplePickerCell<A, B, C> : _PickerCell<Tuple3<A, B, C>> where A: Equ
 
                 let b: B = pickerRow.secondOptions(a).contains(value.b) ? value.b : pickerRow.secondOptions(a)[0]
                 let c: C = pickerRow.thirdOptions(a, b).contains(value.c) ? value.c : pickerRow.thirdOptions(a, b)[0]
+                pickerRow.value = Tuple3(a: a, b: b, c: c)
                 pickerView.reloadComponent(1)
                 pickerView.reloadComponent(2)
-                pickerRow.value = Tuple3(a: a, b: b, c: c)
                 if b != value.b {
                     pickerView.selectRow(0, inComponent: 1, animated: true)
                 }